hw6 refactor: some arch changes
