+++ currency-rates/internal/app/app.go
@@ -8,7 +8,7 @@ import (
 	"github.com/RinaDish/currency-rates/internal/handlers"
 	"github.com/RinaDish/currency-rates/internal/repo"
 	"github.com/RinaDish/currency-rates/internal/routers"
-	"github.com/RinaDish/currency-rates/internal/scheduler"
+	"github.com/RinaDish/currency-rates/internal/workers"
 	"github.com/RinaDish/currency-rates/internal/services"
 	"github.com/RinaDish/currency-rates/tools"
 	"github.com/go-co-op/gocron/v2"
@@ -22,7 +22,7 @@ type App struct {
 	subscriptionHandler handlers.SubscribeHandler
 	ratesHandler handlers.RateHandler
 	subscriptionService services.SubscriptionService
-	subscriptionCron  scheduler.Cron
+	subscriptionCron  workers.Cron
 	router routers.Router
 	db *gorm.DB
 }
@@ -50,7 +50,7 @@ func NewApp(c Config, logger tools.Logger, ctx context.Context) (*App, error) {
 	ratesHandler := handlers.NewRateHandler(logger, rateService)
 	subscriptionHandler := handlers.NewSubscribeHandler(logger, adminRepository)
 
-	cron := scheduler.NewCron(logger)
+	cron := workers.NewCron(logger)
 	task := gocron.NewTask(subscriptionService.NotifySubscribers, ctx)
 	
 	cron.RegisterTask("0 2 * * *", task)
