+++ currency-rates/internal/repo/subsctiption.go
@@ -2,23 +2,20 @@ package repo
 
 import (
 	"context"
+
+	"github.com/RinaDish/currency-rates/internal/services"
 )
 
-type Email struct {
-	ID    int    `json:"id" gorm:"id"`
-	Email string `json:"email" gorm:"email"`
-}
-
 func (repo *Repository) SetEmail(ctx context.Context, email string) error {
-	e := &Email{
+	e := &services.Email{
 		Email: email,
 	}
 
 	return repo.DB.Table("emails").Create(e).Error
 }
 
-func (repo *Repository) GetEmails(ctx context.Context) ([]Email, error) {
-	result := make([]Email, 0)
+func (repo *Repository) GetEmails(ctx context.Context) ([]services.Email, error) {
+	result := make([]services.Email, 0)
 	err := repo.DB.Table("emails").Find(&result).Error
 	
 	return result, err
